Add support for UNC Paths
Added support for UNC server paths. This was done by changing the filesystem on windows to treat all Root Local Device paths as a directory in the root folder, as well as the namespaced UNC path `\\?\UNC`.
Also added some things to allow emulating a win32 fs on *nix environments.
Had to fix some bugs with Node's own handling of UNC paths...